pyproject.toml
[tool.poetry]
name = "bookmark"
version = "0.1.0"
description = ""
authors = ["Spaceocean"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.14.1"
flask-login = "^0.6.3"
flask-session = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
